Refactor cell and grid movement logic with energy updates (Game Executing ok)
- 🛠️ Updated `.uasset` files with new versions and sizes.
- 🔄 Modified `MoveCell` to accept two boolean parameters for better control.
- ⚡ Updated `In_De_creaseCellEnergy` to use a float for energy management.
- 💡 Added `ShareTheEnergy` function for energy sharing during cell division.
- 🔋 Introduced `MoveCellEvent` in `UCPP_AC_Cell_Life` to adjust energy consumption.
- 🚀 Added `UCPP_AC_Grid_Movement` class for managing grid movement.
- 📊 Enhanced energy management properties in `CPP_DA_GameSettings`.
- 🎉 Improved event delegate system for handling cell movement events.

-GAAN

+++ TFM_Prototype/Source/TFM_Prototype/Public/Actors/Cell/CPP_DA_CellType.h
@@ -104,17 +104,17 @@ public:/*Properties*/
 
 	UPROPERTY(EditInstanceOnly, BlueprintReadOnly, Category = "General|Properties|Energy",
 		meta = (ToolTip = "", ClampMin = 0, Units = "Percent"))
-	int MaxEnergy{ 100 };
+	float MaxEnergy{ 100 };
 	UPROPERTY(EditInstanceOnly, BlueprintReadOnly, Category = "General|Properties|Energy",
 		meta = (ToolTip = "Amount of decrease for each time step", ClampMin = 0))
-	int EnergyDecreaseRate{ 1 };
+	float EnergyDecreaseRate{ 1 };
 
 	UPROPERTY(EditInstanceOnly, BlueprintReadOnly, Category = "General|Properties|Energy")
 	bool bCanTransferEnergy{ false };
 
 	UPROPERTY(EditInstanceOnly, BlueprintReadOnly, Category = "General|Properties|Energy",
 		meta = (EditCondition = "bCanTransferEnergy", ClampMin = 0, ClampMax = 100, Units = "Percent"))
-	int EnergyTransferCapacity{ 100 };
+	float EnergyTransferCapacity{ 100 };
